requirements.txt
aiogram==3.15.0
asyncssh==2.22.0
aiosqlite==0.20.0
python-dotenv==1.0.0
apscheduler==3.10.4
aiohttp==3.12.15
cryptography>=46.0.0